Add tests for task models

+++ task/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from todolist.models import ToDoList
 from custom_user.models import CustomUser
-from datetime import datetime
+from datetime import date
 from django.http import HttpResponse
 
 
@@ -27,52 +27,35 @@ class Task(models.Model):
             return None
 
     @classmethod
-    def get_all(cls):
-        try:
-            task = Task.objects.all()
-            return task
-        except Task.DoesNotExist:
-            return None
-
-    @classmethod
     def find_all_for_list(cls, list_id):
         tasks = Task.objects.filter(list_id=list_id)
         return tasks
 
     @classmethod
-    def create(cls, title: str, description: str, deadline: datetime, user_id, task_id):
+    def create(cls, title: str, description: str, deadline: date, user_id, list_id):
         task = Task(title=title, description=description, deadline=deadline)
-        user = CustomUser.find_by_id(user_id)
+        user = CustomUser.get_by_id(user_id)
         task.user_id = user
-        list = ToDoList.get_by_id(task_id)
+        list = ToDoList.get_by_id(list_id)
         task.list_id = list
         task.save()
         return task
 
-    def update(self, new_title: str, new_description: str, is_completed: bool, new_deadline: datetime, user_id: int, list_id: int):
-        self.title = new_title
-        self.description = new_description
-        self.deadline = new_deadline
+    def update(self, title: str, description: str, is_completed: bool, deadline: date, user_id: int, list_id: int):
+        self.title = title
+        self.description = description
+        self.deadline = deadline
         self.is_completed = is_completed
-        self.user_id = CustomUser.find_by_id(user_id)
+        self.user_id = CustomUser.get_by_id(user_id)
         self.list_id = ToDoList.get_by_id(list_id)
         self.save()
-        return self
+        return self.find_by_id(self.pk)
 
     @classmethod
-    def remove(cls, user_id):
-        try:
-            task = Task.objects.get(user_id=user_id)
+    def remove(cls, task_id):
+        task = Task.find_by_id(task_id)
+        if task:
             task.delete()
-        except Task.DoesNotExist:
-            return False
-        return HttpResponse("Task was deleted.")
-
-    @classmethod
-    def remove_all(cls):
-        try:
-            for el in Task.get_all():
-                Task.remove(el.user_id)
-        except Task.DoesNotExist:
-            return False
-        return HttpResponse("All tasks deleted.")
+            return HttpResponse("Task was deleted.")
+        else:
+                return False
